Endpoint de ganancias por dia

+++ database/migrations/2021_08_15_215723_create_orders_table.php
@@ -15,7 +15,7 @@ class CreateOrdersTable extends Migration
     {
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
-            $table->integer('table_id')->unsigned()->index()->nullable();
+            $table->bigInteger('table_id')->unsigned()->index()->nullable();
             $table->foreign('table_id')->references('id')->on('tables');
             $table->text("comments")->nullable();
             $table->enum('state', ['CLOSED', 'OPEN'])->default('OPEN');
