+++ database/migrations/2021_08_15_222030_item_order.php
@@ -15,8 +15,8 @@ class ItemOrder extends Migration
     {
         if (!Schema::hasTable('item_order')) {
             Schema::create('item_order', function(Blueprint $table) {
-                $table->integer('item_id')->unsigned()->index();
-                $table->integer('order_id')->unsigned()->index();
+                $table->bigInteger('item_id')->unsigned()->index();
+                $table->bigInteger('order_id')->unsigned()->index();
                 $table->foreign('item_id')->references('id')->on('items');
                 $table->foreign('order_id')->references('id')->on('orders');
                 $table->integer('items_amount');
